+++ internal/service/commit/server.go
@@ -4,7 +4,7 @@ import pb "gitlab.com/gitlab-org/gitaly-proto/go"
 
 type server struct{}
 
-// NewServer creates a new instance of a grpc CommitServer
+// NewServer creates a new instance of a grpc CommitServiceServer
 func NewServer() pb.CommitServiceServer {
 	return &server{}
 }