+++ internal/service/ref/server.go
@@ -2,13 +2,9 @@ package ref
 
 import pb "gitlab.com/gitlab-org/gitaly-proto/go"
 
-const maxMsgSize = 1024
-
-type server struct {
-	MaxMsgSize int
-}
+type server struct{}
 
 // NewServer creates a new instance of a grpc RefServer
 func NewServer() pb.RefServiceServer {
-	return &server{MaxMsgSize: maxMsgSize}
+	return &server{}
 }