+++ internal/testhelper/testhelper.go
@@ -1,6 +1,7 @@
 package testhelper
 
 import (
+	"bytes"
 	"io"
 	"io/ioutil"
 	"os"
@@ -107,4 +108,34 @@ func MustRunCommand(t *testing.T, stdin io.Reader, name string, args ...string) 
 	}
 
 	return output
+}
+
+// AuthorsEqual tests if two `CommitAuthor`s are equal
+func AuthorsEqual(a *pb.CommitAuthor, b *pb.CommitAuthor) bool {
+	return bytes.Equal(a.Name, b.Name) &&
+		bytes.Equal(a.Email, b.Email) &&
+		a.Date.Seconds == b.Date.Seconds
+}
+
+// CommitsEqual tests if two `GitCommit`s are equal
+func CommitsEqual(a *pb.GitCommit, b *pb.GitCommit) bool {
+	return a.Id == b.Id &&
+		bytes.Equal(a.Subject, b.Subject) &&
+		AuthorsEqual(a.Author, b.Author) &&
+		AuthorsEqual(a.Committer, b.Committer)
+}
+
+// FindLocalBranchCommitAuthorsEqual tests if two `FindLocalBranchCommitAuthor`s are equal
+func FindLocalBranchCommitAuthorsEqual(a *pb.FindLocalBranchCommitAuthor, b *pb.FindLocalBranchCommitAuthor) bool {
+	return bytes.Equal(a.Name, b.Name) &&
+		bytes.Equal(a.Email, b.Email) &&
+		a.Date.Seconds == b.Date.Seconds
+}
+
+// FindLocalBranchResponsesEqual tests if two `FindLocalBranchResponse`s are equal
+func FindLocalBranchResponsesEqual(a *pb.FindLocalBranchResponse, b *pb.FindLocalBranchResponse) bool {
+	return a.CommitId == b.CommitId &&
+		bytes.Equal(a.CommitSubject, b.CommitSubject) &&
+		FindLocalBranchCommitAuthorsEqual(a.CommitAuthor, b.CommitAuthor) &&
+		FindLocalBranchCommitAuthorsEqual(a.CommitCommitter, b.CommitCommitter)
 }