+++ CHANGELOG.md
@@ -4,6 +4,8 @@ UNRELEASED
 
 - Ensure that sub-processes inherit TZ environment variable
   https://gitlab.com/gitlab-org/gitaly/merge_requests/201
+- Implement CommitService::CommitsBetween
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/197
 
 v0.14.0